data_ingestion/producer/requirements.txt
pandas
kafka-python
pyarrow
requests
tqdm